requirements.txt
cassandra-driver==3.3.0
Flask==0.10.1
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
six==1.10.0
Werkzeug==0.11.9
wheel==0.24.0
